backend/requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
fonttools==4.61.1
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pillow==12.1.0
pyparsing==3.3.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.18.0
python-dateutil==2.9.0.post0
reportlab==4.4.9
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
